implemented poolshare tracking by tracking xtoken transfers
- moved template methods for xtoken from xtokenwrapper.ts to xtoken.ts
- no longer tracking joinpool/exitpool events
- moved transfer event handling to xtoken.ts
