+++ src/mappings/xtokenwrapper.ts
@@ -1,9 +1,8 @@
 import { RegisterToken } from '../types/XTokenWrapper/XTokenWrapper'
 import { Token, XToken } from '../types/schema'
 import { ERC20 } from '../types/XTokenWrapper/ERC20'
-import { Paused, Unpaused } from '../types/templates/XToken/XToken'
 import { XToken as XTokenAbi } from '../types/templates'
-import { Address } from '@graphprotocol/graph-ts'
+import { log } from '@graphprotocol/graph-ts'
 
 const DEFAULT_DECIMALS = 18
 
@@ -48,21 +47,5 @@ export function handleRegisterToken(event: RegisterToken): void {
     XTokenAbi.create(xTokenIdAddress)
 
     token.save()
-    xToken.save()
-}
-
-export function handlePaused(event: Paused): void {
-    changeXTokenState(event.address, true)
-}
-
-export function handleUnpaused(event: Unpaused): void {
-    changeXTokenState(event.address, false)
-}
-
-function changeXTokenState(xTokenAddress: Address, xTokenState: boolean): void{
-    let xTokenId = xTokenAddress.toHexString()
-    let xToken = XToken.load(xTokenId)
-
-    xToken.paused = xTokenState
     xToken.save()
 }
